+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:test_provider/provider/counter_provider.dart';
-import 'package:test_provider/second_page.dart';
-import 'package:test_provider/third_page.dart';
+import 'package:test_provider/provider/second_page.dart';
+import 'package:test_provider/provider/third_page.dart';
 
 void main() {
   runApp(const MyApp());
