+++ lib/models/Company.dart
@@ -1,0 +1,23 @@
+class Company {
+  Company({
+    this.name,
+    this.catchPhrase,
+    this.bs,
+  });
+
+  String? name;
+  String? catchPhrase;
+  String? bs;
+
+  factory Company.fromJson(Map<String, String> json) => Company(
+        name: json["name"],
+        catchPhrase: json["catchPhrase"],
+        bs: json["bs"],
+      );
+
+  Map<String, Object?> toJson() => {
+        "name": name,
+        "catchPhrase": catchPhrase,
+        "bs": bs,
+      };
+}
