+++ lib/models/user.dart
@@ -1,10 +1,5 @@
-import 'dart:convert';
-
-List<Users> usersFromJson(String str) =>
-    List<Users>.from(json.decode(str).map((x) => Users.fromJson(x)));
-
-String usersToJson(List<Users> data) =>
-    json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
+import 'Address.dart';
+import 'Company.dart';
 
 class Users {
   Users({
@@ -47,81 +42,5 @@ class Users {
         "phone": phone,
         "website": website,
         "company": company?.toJson(),
-      };
-}
-
-class Address {
-  Address({
-    this.street,
-    this.suite,
-    this.city,
-    this.zipcode,
-    this.geo,
-  });
-
-  String? street;
-  String? suite;
-  String? city;
-  String? zipcode;
-  Geo? geo;
-
-  factory Address.fromJson(Map<String, dynamic> json) => Address(
-        street: json["street"],
-        suite: json["suite"],
-        city: json["city"],
-        zipcode: json["zipcode"],
-        geo: Geo.fromJson(json["geo"]),
-      );
-
-  Map<String, dynamic> toJson() => {
-        "street": street,
-        "suite": suite,
-        "city": city,
-        "zipcode": zipcode,
-        "geo": geo?.toJson(),
-      };
-}
-
-class Geo {
-  Geo({
-    this.lat,
-    this.lng,
-  });
-
-  String? lat;
-  String? lng;
-
-  factory Geo.fromJson(Map<String, dynamic> json) => Geo(
-        lat: json["lat"],
-        lng: json["lng"],
-      );
-
-  Map<String, dynamic> toJson() => {
-        "lat": lat,
-        "lng": lng,
-      };
-}
-
-class Company {
-  Company({
-    this.name,
-    this.catchPhrase,
-    this.bs,
-  });
-
-  String? name;
-  String? catchPhrase;
-  String? bs;
-
-  factory Company.fromJson(Map<String, dynamic> json) => Company(
-        name: json["name"],
-        catchPhrase: json["catchPhrase"],
-        bs: json["bs"],
-      );
-
-  Map<String, dynamic> toJson() => {
-        "name": name,
-        "catchPhrase": catchPhrase,
-        "bs": bs,
       };
 }
