+++ lib/repository/repo.dart
@@ -1,0 +1,9 @@
+import 'dart:convert';
+
+import 'package:test_provider/models/user.dart';
+
+List<Users> usersFromJson(String str) =>
+    List<Users>.from(json.decode(str).map((x) => Users.fromJson(x)));
+
+String usersToJson(List<Users> data) =>
+    json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
