+++ lib/service/user_service.dart
@@ -1,14 +1,14 @@
-import '../models/user.dart';
+import 'package:test_provider/models/user.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert' as convert;
 import 'dart:io';
 
+const _baseUrl = 'https://jsonplaceholder.typicode.com';
+
 class ApiService {
   Future<List<Users>?> fetchHero() async {
-    final response =
-        await http.get(Uri.parse('https://jsonplaceholder.typicode.com/users'));
+    final response = await http.get(Uri.parse('$_baseUrl/users'));
     var jsonResponse = convert.jsonDecode(response.body) as List;
-    print(jsonResponse);
     if (response.statusCode == 200) {
       return jsonResponse.map((e) => Users.fromJson(e)).toList();
     }
