+++ lib/widgets/users_list_models.dart
@@ -1,14 +1,14 @@
 import 'package:elementary/elementary.dart';
 
-import '../models/user.dart';
-import '../service/user_service.dart';
+import 'package:test_provider/models/user.dart';
+import 'package:test_provider/service/user_service.dart';
 
 class UsersListModel extends ElementaryModel {
-  final ApiService? _apiService;
+  final ApiService _apiService;
 
   UsersListModel(this._apiService);
 
   Future<List<Users>?> getUser() async {
-    return _apiService?.fetchHero();
+    return _apiService.fetchHero();
   }
 }
