+++ pubspec.yaml
@@ -29,8 +29,10 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  elementary: ^1.2.0
+  mocktail: ^0.3.0
   provider: ^6.0.2
-  get_it: ^7.2.0
+  http: any
 
 
   # The following adds the Cupertino Icons font to your application.
